pyproject.toml
[project]
name = "llm-retrieval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.14",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appnope==0.1.4",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "authlib==1.6.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.7.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "chromadb==1.0.15",
    "click==8.2.1",
    "coloredlogs==15.0.1",
    "comm==0.2.2",
    "cryptography==45.0.5",
    "cyclopts==3.22.2",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "docstring-parser==0.16",
    "docutils==0.21.2",
    "durationpy==0.10",
    "email-validator==2.2.0",
    "exceptiongroup==1.3.0",
    "executing==2.2.0",
    "fastmcp==2.10.5",
    "filelock==3.18.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.7.0",
    "fsspec==2025.5.1",
    "google-auth==2.40.3",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.73.1",
    "h11==0.16.0",
    "hf-xet==1.1.5",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.33.4",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "ipykernel==6.29.5",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "kubernetes==33.1.0",
    "langchain==0.3.26",
    "langchain-chroma==0.2.4",
    "langchain-community==0.3.27",
    "langchain-core==0.3.68",
    "langchain-mcp-adapters==0.1.9",
    "langchain-openai==0.3.27",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.5.2",
    "langgraph-checkpoint==2.1.0",
    "langgraph-prebuilt==0.5.2",
    "langgraph-sdk==0.1.72",
    "langsmith==0.4.5",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.1.7",
    "mcp[cli]==1.11.0",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "mpmath==1.3.0",
    "multidict==6.6.3",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "numpy==2.3.1",
    "oauthlib==3.3.1",
    "ollama>=0.5.1",
    "onnxruntime==1.22.1",
    "openai==1.95.1",
    "openapi-pydantic==0.5.1",
    "opentelemetry-api==1.35.0",
    "opentelemetry-exporter-otlp-proto-common==1.35.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.35.0",
    "opentelemetry-proto==1.35.0",
    "opentelemetry-sdk==1.35.0",
    "opentelemetry-semantic-conventions==0.56b0",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "overrides==7.7.0",
    "packaging==24.2",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "pip==24.2",
    "platformdirs==4.3.8",
    "posthog==5.4.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.1",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pymupdf==1.26.3",
    "pyperclip==1.9.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "pyzmq==27.0.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "rich-rst==1.3.1",
    "rpds-py==0.26.0",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "scikit-learn==1.7.0",
    "scipy==1.16.0",
    "sentence-transformers==5.0.0",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sse-starlette==2.4.1",
    "stack-data==0.6.3",
    "starlette==0.47.1",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "tokenizers==0.21.2",
    "torch==2.7.1",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.53.2",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.23.0",
]
